+++ Forms/MainForm.cs
@@ -16,7 +16,6 @@ using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.Diagnostics;
 using System.Security.Cryptography;
-using Newtonsoft.Json.Schema.Generation;
 using Newtonsoft.Json.Schema;
 
 namespace Matixs_Mod_Installer
@@ -84,8 +83,6 @@ namespace Matixs_Mod_Installer
             Memory.loadedModpacks.Clear();
             pnlLoadingModpacks.Visible = true;
             int counter = 0;
-            JSchemaGenerator generator = new JSchemaGenerator();
-            JSchema schema = generator.Generate(typeof(ModpackList));
             List<string> loadedUIDs = new List<string>();
             using (WebClient webClient = new WebClient())
             {
@@ -94,12 +91,9 @@ namespace Matixs_Mod_Installer
                 {
                     _log.Info("Loading Modpacks from Source: " + source);
                     string result = await webClient.DownloadStringTaskAsync(source);
-                    if (!JObject.Parse(result).IsValid(schema))
-                    {
-                        _log.Warn("Modpack Source not valid: " + source);
-                        continue;
-                    }
+
                     ModpackList modpacks = JsonConvert.DeserializeObject<ModpackList>(result);
+
                     _log.Debug("Modpacks Provider: " + modpacks.Provider + "(" + modpacks.Website + ")");
                     foreach (Modpack modpack in modpacks.Modpacks)
                     {