+++ changelog.md
@@ -10,7 +10,13 @@
 - [X] Dialog which let user select Minecraft Launcher Location
 
 # v0.3.0.20327
-- [X] Modpack Source Verification using .Net Scheme
+- [X] Modpack Source Verification using .Net Schema
 - [X] Run Updater as Admin
 - [X] Force Update function
 - [X] Add "Minecraft Launcher" folder to installation check
+
+# v0.4.0.20329
+- [X] When updating, restart app as admin with argument "--update-initialization" before unpacking updater to avoid permission errors
+- [X] [Bug] Close options menu when clicking on save
+- [X] Removing .Net Schema because of rate limit
+- [X] Updater redesign